backend-original/requirements.txt
fastapi 
uvicorn 
jinja2 
python-multipart
python-dotenv
firebase-admin
google-cloud-firestore
opencv-python
pytesseract
pycountry
pillow
numpy
